backend/requirements.txt
Django==3.2.12
djangorestframework==3.14.0
djangorestframework-simplejwt== 5.2.2
-U drf-yasg==1.21.7
